[plugin.video.news12] 1.0.2
Added entrypoint, resolverurl cache, tweaked pagination.
