+++ plugin.video.news12/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<addon id="plugin.video.news12" version="1.0.1" name="News12" provider-name="Lunatixz">
+<addon id="plugin.video.news12" version="1.0.2" name="News12" provider-name="Lunatixz">
     <requires>
         <import addon="xbmc.python" version="2.25.0"/>
         <import addon="script.module.simplecache" version="1.0.0"/>
@@ -17,7 +17,7 @@
         <source>https://github.com/Lunatixz/XBMC_Addons/tree/master/plugin.video.news12</source>
         <website>http://news12.com</website>
         <forum>https://forum.kodi.tv/showthread.php?tid=319396</forum>
-        <news>1.0.1[CR]- Fixed Video Quality Error</news>
+        <news>[CR]v1.0.2[CR]- Improved browsing speed.[CR]v.1.0.1[CR]- Fixed Video Quality Error</news>
         <assets>
             <icon>resources/images/icon.png</icon>
             <fanart>resources/images/fanart.jpg</fanart>
